+++ ewebsock/src/native_tungstenite.rs
@@ -1,7 +1,11 @@
 #![allow(deprecated)] // TODO(emilk): Remove when we update tungstenite
 
-use crate::{EventHandler, Result, WsEvent, WsMessage};
-use std::sync::mpsc::{Receiver, TryRecvError};
+use std::{
+    ops::ControlFlow,
+    sync::mpsc::{Receiver, TryRecvError},
+};
+
+use crate::{EventHandler, Options, Result, WsEvent, WsMessage};
 
 /// This is how you send [`WsMessage`]s to the server.
 ///
@@ -12,9 +16,7 @@ pub struct WsSender {
 
 impl Drop for WsSender {
     fn drop(&mut self) {
-        if let Err(err) = self.close() {
-            log::warn!("Failed to close web-socket: {err:?}");
-        }
+        self.close();
     }
 }
 
@@ -28,28 +30,28 @@ impl WsSender {
         }
     }
 
-    /// Close the conenction.
+    /// Close the connection.
     ///
     /// This is called automatically when the sender is dropped.
-    pub fn close(&mut self) -> Result<()> {
+    pub fn close(&mut self) {
         if self.tx.is_some() {
             log::debug!("Closing WebSocket");
         }
         self.tx = None;
-        Ok(())
     }
 
     /// Forget about this sender without closing the connection.
     pub fn forget(mut self) {
+        #[allow(clippy::mem_forget)] // intentional
         std::mem::forget(self.tx.take());
     }
 }
 
-pub(crate) fn ws_receive_impl(url: String, on_event: EventHandler) -> Result<()> {
+pub(crate) fn ws_receive_impl(url: String, options: Options, on_event: EventHandler) -> Result<()> {
     std::thread::Builder::new()
         .name("ewebsock".to_owned())
         .spawn(move || {
-            if let Err(err) = ws_receiver_blocking(&url, &on_event) {
+            if let Err(err) = ws_receiver_blocking(&url, options, &on_event) {
                 on_event(WsEvent::Error(err));
             } else {
                 log::debug!("WebSocket connection closed.");
@@ -62,19 +64,23 @@ pub(crate) fn ws_receive_impl(url: String, on_event: EventHandler) -> Result<()>
 
 /// Connect and call the given event handler on each received event.
 ///
-/// Blocking version of [`ws_receive`], only avilable on native.
+/// Blocking version of [`ws_receive`], only available on native.
 ///
 /// # Errors
-/// * Any connection failures
-pub fn ws_receiver_blocking(url: &str, on_event: &EventHandler) -> Result<()> {
+/// All errors are returned to the caller, and NOT reported via `on_event`.
+pub fn ws_receiver_blocking(url: &str, options: Options, on_event: &EventHandler) -> Result<()> {
     let uri: http::Uri = url.parse().unwrap();
     let builder = tungstenite::ClientRequestBuilder::new(uri).with_header("Origin", url);
-    let (mut socket, response) = match tungstenite::connect(builder) {
-        Ok(result) => result,
-        Err(err) => {
-            return Err(format!("Connect: {err}"));
-        }
-    };
+    let config = tungstenite::protocol::WebSocketConfig::from(options);
+    let max_redirects = 3; // tungstenite default
+
+    let (mut socket, response) =
+        match tungstenite::client::connect_with_config(builder, Some(config), max_redirects) {
+            Ok(result) => result,
+            Err(err) => {
+                return Err(format!("Connect: {err}"));
+            }
+        };
 
     log::debug!("WebSocket HTTP response code: {}", response.status());
     log::trace!(
@@ -82,47 +88,63 @@ pub fn ws_receiver_blocking(url: &str, on_event: &EventHandler) -> Result<()> {
         response.headers()
     );
 
-    on_event(WsEvent::Opened);
+    let control = on_event(WsEvent::Opened);
+    if control.is_break() {
+        log::trace!("Closing connection due to Break");
+        return socket
+            .close(None)
+            .map_err(|err| format!("Failed to close connection: {err}"));
+    }
 
     loop {
-        match socket.read_message() {
+        let control = match socket.read_message() {
             Ok(incoming_msg) => match incoming_msg {
                 tungstenite::protocol::Message::Text(text) => {
-                    on_event(WsEvent::Message(WsMessage::Text(text)));
+                    on_event(WsEvent::Message(WsMessage::Text(text)))
                 }
                 tungstenite::protocol::Message::Binary(data) => {
-                    on_event(WsEvent::Message(WsMessage::Binary(data)));
+                    on_event(WsEvent::Message(WsMessage::Binary(data)))
                 }
                 tungstenite::protocol::Message::Ping(data) => {
-                    on_event(WsEvent::Message(WsMessage::Ping(data)));
+                    on_event(WsEvent::Message(WsMessage::Ping(data)))
                 }
                 tungstenite::protocol::Message::Pong(data) => {
-                    on_event(WsEvent::Message(WsMessage::Pong(data)));
+                    on_event(WsEvent::Message(WsMessage::Pong(data)))
                 }
                 tungstenite::protocol::Message::Close(close) => {
                     on_event(WsEvent::Closed);
                     log::debug!("WebSocket close received: {close:?}");
                     return Ok(());
                 }
-                tungstenite::protocol::Message::Frame(_) => {}
+                tungstenite::protocol::Message::Frame(_) => ControlFlow::Continue(()),
             },
             Err(err) => {
-                let msg = format!("read: {err}");
-                return Err(msg);
+                return Err(format!("read: {err}"));
             }
+        };
+
+        if control.is_break() {
+            log::trace!("Closing connection due to Break");
+            return socket
+                .close(None)
+                .map_err(|err| format!("Failed to close connection: {err}"));
         }
 
         std::thread::sleep(std::time::Duration::from_millis(10));
     }
 }
 
-pub(crate) fn ws_connect_impl(url: String, on_event: EventHandler) -> Result<WsSender> {
+pub(crate) fn ws_connect_impl(
+    url: String,
+    options: Options,
+    on_event: EventHandler,
+) -> Result<WsSender> {
     let (tx, rx) = std::sync::mpsc::channel();
 
     std::thread::Builder::new()
         .name("ewebsock".to_owned())
         .spawn(move || {
-            if let Err(err) = ws_connect_blocking(&url, &on_event, &rx) {
+            if let Err(err) = ws_connect_blocking(&url, options, &on_event, &rx) {
                 on_event(WsEvent::Error(err));
             } else {
                 log::debug!("WebSocket connection closed.");
@@ -135,23 +157,27 @@ pub(crate) fn ws_connect_impl(url: String, on_event: EventHandler) -> Result<WsS
 
 /// Connect and call the given event handler on each received event.
 ///
-/// This is a blocking variant of [`ws_connect`], only availble on native.
+/// This is a blocking variant of [`ws_connect`], only available on native.
 ///
 /// # Errors
-/// * Any connection failures
+/// All errors are returned to the caller, and NOT reported via `on_event`.
 pub fn ws_connect_blocking(
     url: &str,
+    options: Options,
     on_event: &EventHandler,
     rx: &Receiver<WsMessage>,
 ) -> Result<()> {
+    let config = tungstenite::protocol::WebSocketConfig::from(options);
+    let max_redirects = 3; // tungstenite default
     let uri: http::Uri = url.parse().unwrap();
     let builder = tungstenite::ClientRequestBuilder::new(uri).with_header("Origin", url);
-    let (mut socket, response) = match tungstenite::connect(builder) {
-        Ok(result) => result,
-        Err(err) => {
-            return Err(format!("Connect: {err}"));
-        }
-    };
+    let (mut socket, response) =
+        match tungstenite::client::connect_with_config(builder, Some(config), max_redirects) {
+            Ok(result) => result,
+            Err(err) => {
+                return Err(format!("Connect: {err}"));
+            }
+        };
 
     log::debug!("WebSocket HTTP response code: {}", response.status());
     log::trace!(
@@ -159,7 +185,13 @@ pub fn ws_connect_blocking(
         response.headers()
     );
 
-    on_event(WsEvent::Opened);
+    let control = on_event(WsEvent::Opened);
+    if control.is_break() {
+        log::trace!("Closing connection due to Break");
+        return socket
+            .close(None)
+            .map_err(|err| format!("Failed to close connection: {err}"));
+    }
 
     match socket.get_mut() {
         tungstenite::stream::MaybeTlsStream::Plain(stream) => stream.set_nonblocking(true),
@@ -203,39 +235,45 @@ pub fn ws_connect_blocking(
             Err(TryRecvError::Empty) => {}
         };
 
-        match socket.read_message() {
+        let control = match socket.read_message() {
             Ok(incoming_msg) => {
                 did_work = true;
                 match incoming_msg {
                     tungstenite::protocol::Message::Text(text) => {
-                        on_event(WsEvent::Message(WsMessage::Text(text)));
+                        on_event(WsEvent::Message(WsMessage::Text(text)))
                     }
                     tungstenite::protocol::Message::Binary(data) => {
-                        on_event(WsEvent::Message(WsMessage::Binary(data)));
+                        on_event(WsEvent::Message(WsMessage::Binary(data)))
                     }
                     tungstenite::protocol::Message::Ping(data) => {
-                        on_event(WsEvent::Message(WsMessage::Ping(data)));
+                        on_event(WsEvent::Message(WsMessage::Ping(data)))
                     }
                     tungstenite::protocol::Message::Pong(data) => {
-                        on_event(WsEvent::Message(WsMessage::Pong(data)));
+                        on_event(WsEvent::Message(WsMessage::Pong(data)))
                     }
                     tungstenite::protocol::Message::Close(close) => {
                         on_event(WsEvent::Closed);
                         log::debug!("Close received: {close:?}");
                         return Ok(());
                     }
-                    tungstenite::protocol::Message::Frame(_) => {}
+                    tungstenite::protocol::Message::Frame(_) => ControlFlow::Continue(()),
                 }
             }
             Err(tungstenite::Error::Io(io_err))
                 if io_err.kind() == std::io::ErrorKind::WouldBlock =>
             {
-                // Ignore
+                ControlFlow::Continue(()) // Ignore
             }
             Err(err) => {
-                let msg = format!("read: {err}");
-                return Err(msg);
+                return Err(format!("read: {err}"));
             }
+        };
+
+        if control.is_break() {
+            log::trace!("Closing connection due to Break");
+            return socket
+                .close(None)
+                .map_err(|err| format!("Failed to close connection: {err}"));
         }
 
         if !did_work {