+++ ewebsock/src/native_tungstenite_tokio.rs
@@ -1,4 +1,6 @@
-use crate::{EventHandler, Result, WsEvent, WsMessage};
+use std::ops::ControlFlow;
+
+use crate::{EventHandler, Options, Result, WsEvent, WsMessage};
 
 /// This is how you send [`WsMessage`]s to the server.
 ///
@@ -9,9 +11,7 @@ pub struct WsSender {
 
 impl Drop for WsSender {
     fn drop(&mut self) {
-        if let Err(err) = self.close() {
-            log::warn!("Failed to close web-socket: {err:?}");
-        }
+        self.close();
     }
 }
 
@@ -25,31 +25,40 @@ impl WsSender {
         }
     }
 
-    /// Close the conenction.
+    /// Close the connection.
     ///
     /// This is called automatically when the sender is dropped.
-    pub fn close(&mut self) -> Result<()> {
+    pub fn close(&mut self) {
         if self.tx.is_some() {
             log::debug!("Closing WebSocket");
         }
         self.tx = None;
-        Ok(())
     }
 
     /// Forget about this sender without closing the connection.
     pub fn forget(mut self) {
+        #[allow(clippy::mem_forget)] // intentional
         std::mem::forget(self.tx.take());
     }
 }
 
 async fn ws_connect_async(
     url: String,
+    options: Options,
     outgoing_messages_stream: impl futures::Stream<Item = WsMessage>,
     on_event: EventHandler,
 ) {
     use futures::StreamExt as _;
 
-    let (ws_stream, _) = match tokio_tungstenite::connect_async(url).await {
+    let config = tungstenite::protocol::WebSocketConfig::from(options);
+    let disable_nagle = false; // God damn everyone who adds negations to the names of their variables
+    let (ws_stream, _response) = match tokio_tungstenite::connect_async_with_config(
+        url,
+        Some(config),
+        disable_nagle,
+    )
+    .await
+    {
         Ok(result) => result,
         Err(err) => {
             on_event(WsEvent::Error(err.to_string()));
@@ -58,7 +67,11 @@ async fn ws_connect_async(
     };
 
     log::info!("WebSocket handshake has been successfully completed");
-    on_event(WsEvent::Opened);
+
+    let control = on_event(WsEvent::Opened);
+    if control.is_break() {
+        log::warn!("ControlFlow::Break not implemented for the tungstenite tokio backend");
+    }
 
     let (write, read) = ws_stream.split();
 
@@ -74,29 +87,28 @@ async fn ws_connect_async(
         .forward(write);
 
     let reader = read.for_each(move |event| {
-        match event {
+        let control = match event {
             Ok(message) => match message {
                 tungstenite::protocol::Message::Text(text) => {
-                    on_event(WsEvent::Message(WsMessage::Text(text)));
+                    on_event(WsEvent::Message(WsMessage::Text(text)))
                 }
                 tungstenite::protocol::Message::Binary(data) => {
-                    on_event(WsEvent::Message(WsMessage::Binary(data)));
+                    on_event(WsEvent::Message(WsMessage::Binary(data)))
                 }
                 tungstenite::protocol::Message::Ping(data) => {
-                    on_event(WsEvent::Message(WsMessage::Ping(data)));
+                    on_event(WsEvent::Message(WsMessage::Ping(data)))
                 }
                 tungstenite::protocol::Message::Pong(data) => {
-                    on_event(WsEvent::Message(WsMessage::Pong(data)));
+                    on_event(WsEvent::Message(WsMessage::Pong(data)))
                 }
-                tungstenite::protocol::Message::Close(_) => {
-                    on_event(WsEvent::Closed);
-                }
-                tungstenite::protocol::Message::Frame(_) => {}
+                tungstenite::protocol::Message::Close(_) => on_event(WsEvent::Closed),
+                tungstenite::protocol::Message::Frame(_) => ControlFlow::Continue(()),
             },
-            Err(err) => {
-                on_event(WsEvent::Error(err.to_string()));
-            }
+            Err(err) => on_event(WsEvent::Error(err.to_string())),
         };
+        if control.is_break() {
+            log::warn!("ControlFlow::Break not implemented for the tungstenite tokio backend");
+        }
         async {}
     });
 
@@ -105,12 +117,16 @@ async fn ws_connect_async(
 }
 
 #[allow(clippy::unnecessary_wraps)]
-pub(crate) fn ws_connect_impl(url: String, on_event: EventHandler) -> Result<WsSender> {
-    Ok(ws_connect_native(url, on_event))
+pub(crate) fn ws_connect_impl(
+    url: String,
+    options: Options,
+    on_event: EventHandler,
+) -> Result<WsSender> {
+    Ok(ws_connect_native(url, options, on_event))
 }
 
-/// Like [`ws_connect`], but cannot fail. Only available on native builds.
-fn ws_connect_native(url: String, on_event: EventHandler) -> WsSender {
+/// Like [`crate::ws_connect`], but cannot fail. Only available on native builds.
+fn ws_connect_native(url: String, options: Options, on_event: EventHandler) -> WsSender {
     let (tx, mut rx) = tokio::sync::mpsc::channel(1000);
 
     let outgoing_messages_stream = async_stream::stream! {
@@ -121,12 +137,12 @@ fn ws_connect_native(url: String, on_event: EventHandler) -> WsSender {
     };
 
     tokio::spawn(async move {
-        ws_connect_async(url.clone(), outgoing_messages_stream, on_event).await;
+        ws_connect_async(url.clone(), options, outgoing_messages_stream, on_event).await;
         log::debug!("WS connection finished.");
     });
     WsSender { tx: Some(tx) }
 }
 
-pub(crate) fn ws_receive_impl(url: String, on_event: EventHandler) -> Result<()> {
-    ws_connect_impl(url, on_event).map(|sender| sender.forget())
+pub(crate) fn ws_receive_impl(url: String, options: Options, on_event: EventHandler) -> Result<()> {
+    ws_connect_impl(url, options, on_event).map(|sender| sender.forget())
 }